refactor: onboarding module
