+++ packages/cryptoplease/integration_test/app_test.dart
@@ -1,5 +1,5 @@
-import 'package:cryptoplease/features/onboarding/presentation/create_profile_screen.dart';
-import 'package:cryptoplease/features/onboarding/presentation/get_started_screen.dart';
+import 'package:cryptoplease/features/onboarding/src/presentation/create_profile_screen.dart';
+import 'package:cryptoplease/features/onboarding/src/presentation/get_started_screen.dart';
 import 'package:cryptoplease/main.dart' as app;
 import 'package:cryptoplease/ui/bottom_button.dart';
 import 'package:cryptoplease/ui/recovery_phrase_text_view.dart';
