+++ packages/cryptoplease/lib/app/screens/authenticated/profile/edit_profile_screen.dart
@@ -4,11 +4,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../core/accounts/bl/accounts_bloc.dart';
-import '../../../../features/onboarding/presentation/components/pick_profile_picture.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../../ui/app_bar.dart';
 import '../../../../ui/loader.dart';
 import '../../../../ui/onboarding_screen.dart';
+import '../../../../ui/profile_image_picker/pick_profile_picture.dart';
 import '../../../../ui/text_field.dart';
 import '../../../../ui/theme.dart';
 
@@ -62,7 +62,7 @@ class _EditProfileScreenState extends State<EditProfileScreen> {
               ),
               children: [
                 CpAppBar(leading: BackButton(onPressed: _closeFlow)),
-                PickProfilePicture(
+                ProfileImagePicker(
                   onChanged: (photo) => setState(() {
                     _photo = photo;
                   }),
