+++ packages/cryptoplease/lib/features/onboarding/src/presentation/no_email_and_password_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-import '../../../l10n/l10n.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/info_widget.dart';
-import '../../../ui/onboarding_screen.dart';
-import '../../../ui/theme.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/app_bar.dart';
+import '../../../../ui/info_widget.dart';
+import '../../../../ui/onboarding_screen.dart';
+import '../../../../ui/theme.dart';
 import 'onboarding_flow_screen.dart';
 
 class NoEmailAndPasswordScreen extends StatelessWidget {
