+++ packages/cryptoplease/lib/features/onboarding/src/presentation/restore_account_screen.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../l10n/l10n.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/colors.dart';
-import '../../../ui/onboarding_screen.dart';
-import '../../../ui/theme.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/app_bar.dart';
+import '../../../../ui/colors.dart';
+import '../../../../ui/onboarding_screen.dart';
+import '../../../../ui/theme.dart';
 import '../bl/onboarding_bloc.dart';
 import 'components/mnemonic_input_formatter.dart';
 import 'onboarding_flow_screen.dart';
