+++ packages/cryptoplease/test/golden/golden_test.dart
@@ -8,8 +8,8 @@ import 'package:cryptoplease/core/conversion_rates/bl/repository.dart';
 import 'package:cryptoplease/core/currency.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/user_preferences.dart';
-import 'package:cryptoplease/features/onboarding/presentation/get_started_screen.dart';
-import 'package:cryptoplease/features/onboarding/presentation/no_email_and_password_screen.dart';
+import 'package:cryptoplease/features/onboarding/src/presentation/get_started_screen.dart';
+import 'package:cryptoplease/features/onboarding/src/presentation/no_email_and_password_screen.dart';
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
